Changes to format, created the board class,  and started the functionality to crawler move() function.
Created the board.cpp and board.h file, I have moved my fillBugsVector() function to the board and then calling it in the main.cpp, I forgot this was in the brief and moved it across. I also have finished the isWayBlocked() functionality , and have started the move function for crawler bugs.
