+++ Board.h
@@ -1,0 +1,18 @@
+//
+// Created by caitl on 18/04/2024.
+//
+
+#ifndef CA2_BUGS_LIFE_BOARD_H
+#define CA2_BUGS_LIFE_BOARD_H
+
+#include "Bug.h"
+#include <vector>
+#include <fstream> //to access file input and output
+
+//class Board{
+//
+//public:
+    void fillBugsVector(std::ifstream &fin, std::vector<Bug *> &bug_vector);
+//};
+
+#endif //CA2_BUGS_LIFE_BOARD_H
