+++ Bug.cpp
@@ -12,22 +12,40 @@ Bug::~Bug() {
 }
 
 //TODO: implement logic for isWayBlocked()
+// the bug board is 10 x 10 so max is so min cell is 0 and max is 9
 bool Bug::isWayBlocked() {
 
     int x = this->position.first;
     int y = this->position.second;
 
-    // (x,y)
+    // (x,y) x is row , y is column
+    // north and south will be y
+    // east and west will be x
 
     // check if bug facing north
-    if (x == North && y == 0) {
+    if (x == 0 && direction == Direction::North) {
+        cout << "bugs way is blocked facing north" << endl;
         return true;
     }
 
-    //check if bug facing east
-
     //check if bug facing south
+    if(x == 9 && direction == South){
+        cout << "bugs way is blocked facing south" << endl;
+        return true;
+    }
 
-    //check if bug facing west
+    // check if bug facing east
+    // ending cell is 9
+    if(direction == East && y == 9){
+        cout << "bugs way is blocked facing east" << endl;
+        return true;
+    }
+
+    // check if bug facing west
+    // first cell is 0
+    if(direction == West &&  y == 0){
+        cout << "bugs way is blocked facing west" << endl;
+        return true;
+    }
 
 }
