+++ Hopper.cpp
@@ -9,8 +9,10 @@ using namespace std;
 Hopper::Hopper(int id, int x, int y, Direction direction1, int size, bool alive, int hopLength) {
 }
 
+// need to check if wayIsBlocked() everytime a bug moves - if it is then move to a random cell
 void Hopper::move() {      //implementation of Hopper move()
     cout << "Moving a hopper bug " << endl;
 
 }
+
