+++ cmake-build-debug/.ninja_log
@@ -1,13 +1,21 @@
 # ninja log v5
-9	2510	7349973116159324	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
-21	2132	7349973112139662	CMakeFiles/CA2_Bugs_Life.dir/Bug.cpp.obj	7b8eb1854ac61b4b
-2510	3099	7349973121814677	CA2_Bugs_Life.exe	29ad3ef805d9b804
-89	2778	7349961559791922	CMakeFiles/CA2_Bugs_Life.dir/Hopper.cpp.obj	4f00c8d9a3cbea1f
-17	2770	7349961559612667	CMakeFiles/CA2_Bugs_Life.dir/Crawler.cpp.obj	2c9472d55ed28825
-12	1238	7349975516785301	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
-1239	1455	7349975518808762	CA2_Bugs_Life.exe	29ad3ef805d9b804
-125	2336	7349990159523462	CMakeFiles/CA2_Bugs_Life.dir/Hopper.cpp.obj	4f00c8d9a3cbea1f
-25	2339	7349990159543540	CMakeFiles/CA2_Bugs_Life.dir/Crawler.cpp.obj	2c9472d55ed28825
-11	2684	7349990163032894	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
-12	1471	7349990235750208	CMakeFiles/CA2_Bugs_Life.dir/Bug.cpp.obj	7b8eb1854ac61b4b
-1473	2046	7349990241329766	CA2_Bugs_Life.exe	29ad3ef805d9b804
+7	1075	7351424238917051	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
+35	3483	7351423715680740	CMakeFiles/CA2_Bugs_Life.dir/Bug.cpp.obj	7b8eb1854ac61b4b
+1076	1229	7351424240356964	CA2_Bugs_Life.exe	29ad3ef805d9b804
+221	3421	7351423714889623	CMakeFiles/CA2_Bugs_Life.dir/Hopper.cpp.obj	4f00c8d9a3cbea1f
+4	3423	7351423714939414	CMakeFiles/CA2_Bugs_Life.dir/Crawler.cpp.obj	2c9472d55ed28825
+145	1550	7353169323652704	CMakeFiles/CA2_Bugs_Life.dir/Board.cpp.obj	10106eb60db024b4
+136	3591	7353169344437850	CMakeFiles/CA2_Bugs_Life.dir/Hopper.cpp.obj	4f00c8d9a3cbea1f
+24	3604	7353169344543213	CMakeFiles/CA2_Bugs_Life.dir/Bug.cpp.obj	7b8eb1854ac61b4b
+38	3608	7353169344552971	CMakeFiles/CA2_Bugs_Life.dir/Crawler.cpp.obj	2c9472d55ed28825
+11	3727	7353169345797273	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
+3727	4416	7353169352567859	CA2_Bugs_Life.exe	40d08f8df09958c3
+6	4907	7353173312266399	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
+13	4952	7353173312929840	CMakeFiles/CA2_Bugs_Life.dir/Board.cpp.obj	10106eb60db024b4
+6	1226	7353173623759314	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
+9	1174	7353173702238555	CMakeFiles/CA2_Bugs_Life.dir/Board.cpp.obj	10106eb60db024b4
+10	4275	7353190241769591	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
+23	4325	7353190242558928	CMakeFiles/CA2_Bugs_Life.dir/Board.cpp.obj	10106eb60db024b4
+4326	5124	7353190250450435	CA2_Bugs_Life.exe	40d08f8df09958c3
+6	1420	7353190535120173	CMakeFiles/CA2_Bugs_Life.dir/main.cpp.obj	61e03b7d9c9590c4
+1421	1661	7353190537345279	CA2_Bugs_Life.exe	40d08f8df09958c3
