+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Apr 16 20:30 GMT Summer Time
+Start testing: Apr 20 13:24 GMT Summer Time
 ----------------------------------------------------------
-End testing: Apr 16 20:30 GMT Summer Time
+End testing: Apr 20 13:24 GMT Summer Time
